+++ demo/src/main/webapp/js/samaxesjs.core.js
@@ -28,3 +28,9 @@ try {
 } catch (e) {
 }
 
+/**
+ * DEFINE_TEST should be false after Closure Compiler executes
+ * @define {boolean}
+ */
+var DEFINE_TEST = true;
+